+++ libs/slu-ipc/src/error.rs
@@ -4,12 +4,10 @@ use thiserror::Error;
 pub enum Error {
     #[error("IO Error: {0}")]
     Io(#[from] std::io::Error),
-    #[error("Error while encoding using bincode: {0}")]
-    BincodeEncode(#[from] bincode::error::EncodeError),
-    #[error("Error while decoding using bincode: {0}")]
-    BincodeDecode(#[from] bincode::error::DecodeError),
     #[error("Service Error: {0}")]
     IpcResponseError(String),
+    #[error("Serde Json Error: {0}")]
+    SerdeJson(#[from] serde_json::Error),
 }
 
 pub type Result<T = ()> = core::result::Result<T, Error>;