+++ src/background/widgets/window_manager/handler.rs
@@ -39,7 +39,7 @@ pub async fn set_app_windows_positions(positions: HashMap<isize, Rect>) -> Resul
     let state = FULL_STATE.load();
     let perf_mode = PERFORMANCE_MODE.load();
     let place_animated =
-        !state.settings.by_widget.wm.animations.enabled || **perf_mode != PerformanceMode::Disabled;
+        state.settings.by_widget.wm.animations.enabled && **perf_mode == PerformanceMode::Disabled;
 
     ServicePipe::request(SvcAction::DeferWindowPositions {
         list,