+++ src/service/hotkeys.rs
@@ -1,5 +1,5 @@
 use seelen_core::state::shortcuts::{SluHotkeyAction, SluShortcutsSettings};
-use slu_ipc::AppIpc;
+use slu_ipc::{messages::AppMessage, AppIpc};
 use win_hotkeys::{error::WHKError, events::KeyboardInputEvent, Hotkey, HotkeyManager, VKey};
 
 use crate::{app_management::kill_seelen_ui_processes, error::Result, log_error};
@@ -39,7 +39,7 @@ pub fn start_app_shortcuts(config: SluShortcutsSettings) -> Result<()> {
 
             if let Some(command) = hotkey_action_to_cli_command(action) {
                 tokio_handle.spawn(async move {
-                    log_error!(AppIpc::send(command).await);
+                    log_error!(AppIpc::send(AppMessage(command)).await);
                 });
             }
         });
@@ -92,11 +92,11 @@ pub async fn stop_shortcut_registration() -> Result<()> {
 }
 
 async fn send_registering_to_app(hotkey: Option<Vec<String>>) -> Result<()> {
-    AppIpc::send(vec![
+    AppIpc::send(AppMessage(vec![
         "popup".to_owned(),
         "internal-set-shortcut".to_owned(),
         serde_json::to_string(&hotkey)?,
-    ])
+    ]))
     .await?;
     Ok(())
 }